+++ apps/web/src/components/quick-start/common/index.ts
@@ -1,0 +1,1 @@
+export { NodeStep } from './NodeStep';
