+++ apps/web/src/components/quick-start/digest-demo-flow/DigestDemoFlow.tsx
@@ -4,11 +4,11 @@ import { Skeleton, useMantineColorScheme } from '@mantine/core';
 import { useResizeObserver } from '@mantine/hooks';
 import ReactFlow, { Edge, Node, useReactFlow } from 'react-flow-renderer';
 
-import { colors } from '../../design-system';
+import { colors } from '../../../design-system';
 import { TriggerNode } from './TriggerNode';
 import { DigestNode } from './DigestNode';
 import { EmailNode } from './EmailNode';
-import { useTemplateFetcher } from '../../api/hooks';
+import { useTemplateFetcher } from '../../../api/hooks';
 import { DigestDemoFlowProvider } from './DigestDemoFlowProvider';
 
 export function DigestDemoFlow({
