+++ apps/web/src/components/quick-start/digest-demo-flow/Indicator.tsx
@@ -1,7 +1,7 @@
 import { useRef } from 'react';
 import styled from '@emotion/styled';
 
-import { colors } from '../../design-system';
+import { colors } from '../../../design-system';
 
 const IndicatorHolder = styled.span<{ isShown }>`
   display: flex;
