+++ apps/web/src/components/quick-start/digest-demo-flow/TriggerNode.tsx
@@ -1,15 +1,15 @@
 import { Handle, Position } from 'react-flow-renderer';
 
-import { NodeStep } from './NodeStep';
-import { TurnOnGradient } from '../../design-system/icons/gradient/TurnOnGradient';
-import { Button } from '../../design-system';
+import { NodeStepWithPopover } from './NodeStepWithPopover';
+import { TurnOnGradient } from '../../../design-system/icons/gradient/TurnOnGradient';
+import { Button } from '../../../design-system';
 import { useDigestDemoFlowContext } from './DigestDemoFlowProvider';
 
 export function TriggerNode({ data, id }: { data: any; id: string }) {
   const { isReadOnly, runTrigger } = useDigestDemoFlowContext();
 
   return (
-    <NodeStep
+    <NodeStepWithPopover
       data={data}
       id={id}
       Icon={TurnOnGradient}
