+++ apps/web/src/components/quick-start/index.ts
@@ -1,0 +1,2 @@
+export * from './digest-demo-flow';
+export * from './in-app-onboarding';
